make get_proxy to be private.
