+++ src/proxy.rs
@@ -663,4 +663,30 @@ mod tests {
         assert_eq!(intercepted_uri(&p, https), target1);
         assert!(p.intercept(&url(other)).is_none());
     }
+
+    #[test]
+    fn test_get_proxies() {
+        // save system setting first.
+        let system_proxy = env::var("http_proxy");
+
+        // remove proxy.
+        env::remove_var("http_proxy");
+        assert_eq!(get_proxies().contains_key("http"), false);
+
+        // the system proxy setting url is invalid.
+        env::set_var("http_proxy", "123465");
+        assert_eq!(get_proxies().contains_key("http"), false);
+
+        // set valid proxy
+        env::set_var("http_proxy", "http://127.0.0.1/");
+        let proxies = get_proxies();
+        let http_target = proxies.get("http").unwrap().as_str();
+
+        assert_eq!(http_target, "http://127.0.0.1/");
+        // reset user setting.
+        match system_proxy {
+            Err(_) => env::remove_var("http_proxy"),
+            Ok(proxy) => env::set_var("http_proxy", proxy)
+        }
+    }
 }
